Import Panacea_Drums v2.1 by Tsunadee
This is the final version of Panacea_Drums.

It was created by Tsunadee from <Panacea>, a European WoW private
server guild. And it was based on ccDrums v2.0 by ccKep.

Thanks to the earlier ccDrums source code import, this repository
now shows exactly what Panacea_Drums changed. But most importantly,
it serves as a way to download the addon in the future!

Enjoy!